Don't blindly transform alphabets to GrowingAlphabets
Previously, we delibrately copied a given input alphabet to a
GrowingAlphabet. This introduced a performance cost for everyone,
even if this feature was not used. With this change, we only copy
the alphabet on the first utilization of this feature. Additionally,
we perform a defensive copy, i.e. if the user already provided a
(potentially optimized) GrowingAlphabet instance, we do not copy any
alphabets at all.