Refactor DOT visualization
The general intent of this commit is to remove DOT-specific rendering code and
establish a rather generic rendering interface in which other visualization
providers (such as JUNG, currently) can be used in a plug and play fashion.
Especially, if one is not interested in rendering at all, no DOT/JUNG dependencies
should be pulled.

Major refactorings are:

* Added serialization/dot artifact
* Added visualization/dot artifact (this is now required as a runtime dependency
  if one wants to use the DOT renderer)
* Removed common/dotutil (has been moved/merged with visualization/dot)
* Generally the term DOTHelper has been replaced with VisualizationHelper