exclude examples module in a release
Previously we skipped deployment of the examples artifact but it was still included in site generation, which leaked e.g. JavaDoc to the public and caused confusion. This commit excludes the examples module when "automatalib.release" is set to "true". This commit also disables running the code-analysis profile on a release and instead runs it during the <preparationGoals> so that in case of an error, the (broken) release is not tagged in the repository.