remove Serializable markup
We never fully supported the semantics of the Serializable interface and never intended to do so. In fact, the old approach failed miserably if any class was involved where we missed to write "implements Serializable". In order to prevent confusion by promising false contracts, implementing this markup interface is removed. Serialization should now be done in user-land via one of the many external (and optimized) serialization frameworks such as FST, XStream, etc.