cleanup {Static,Dynamic}{State,Node}Mappings
Accessing potential null-values is an inherent data-flow problem. By annotating the result type as @Nullable, we force all consumers to check for null, even in those cases where previous dataflow ensures non-nullability. The alternative would be to additionally track which items have been set and throw an IllegalArgumentException when not able to return a non-null value. However, since this will cause an exception as well, we might as well just let the code crash with a NullPointerException. Yes, it is not fail-fast, but the noise and overhead introduced by explicitly checking nullability does not justify the gains.